requirements.txt
Django==3.2.16
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
drf-spectacular==0.26.5